Added some URLhelpers
