+++ Marvel Heroes/URLHelper/UrlHost.swift
@@ -1,0 +1,12 @@
+//
+//  UrlHost.swift
+//  Marvel Heroes
+//
+//  Created by Alexander Tsygankov on 28.09.2021.
+//
+
+import Foundation
+
+enum UrlHost: String {
+    case marverGateway = "gateway.marvel.com"
+}
