+++ Marvel Heroes/URLHelper/UrlPath.swift
@@ -1,0 +1,13 @@
+//
+//  UrlPath.swift
+//  Marvel Heroes
+//
+//  Created by Alexander Tsygankov on 28.09.2021.
+//
+
+import Foundation
+
+enum UrlPath: String {
+    case characteresUrl = "/v1/public/characters"
+    case comicsUrl = "/v1/public/comics"
+}
