+++ Marvel Heroes/URLHelper/UrlSchema.swift
@@ -1,0 +1,12 @@
+//
+//  UrlSchema.swift
+//  Marvel Heroes
+//
+//  Created by Alexander Tsygankov on 28.09.2021.
+//
+
+import Foundation
+
+enum UrlSchema: String {
+    case https = "https"
+}
